+++ src/networking/index.ts
@@ -1,3 +1,4 @@
 export * from './restClient';
+export * from './types';
 import RestApi from './RestApi';
 export { RestApi };